[refactor] #133 Content Live, Test 분리
